third_party/WebKit/LayoutTests/fast/css/border-image-value-grammar.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description("Support consecutive slash operators for border-image");

var element = document.createElement('div');
var style = element.style;
style.borderImage = '3 / / 2';
shouldBeEqualToString("style.getPropertyValue('border-image-slice')", "3");
shouldBeEqualToString("style.getPropertyValue('border-image-width')", "1");
shouldBeEqualToString("style.getPropertyValue('border-image-outset')", "2");

style.borderImage = '4 / 5';
shouldBeEqualToString("style.getPropertyValue('border-image-slice')", "4");
shouldBeEqualToString("style.getPropertyValue('border-image-width')", "5");
shouldBeEqualToString("style.getPropertyValue('border-image-outset')", "0");

style.borderImage = '1 / 2 / 3';
shouldBeEqualToString("style.getPropertyValue('border-image-slice')", "1");
shouldBeEqualToString("style.getPropertyValue('border-image-width')", "2");
shouldBeEqualToString("style.getPropertyValue('border-image-outset')", "3");
</script>
