third_party/WebKit/LayoutTests/fast/css/border-image-value-grammar-expected.txt

Support consecutive slash operators for border-image

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.getPropertyValue('border-image-slice') is "3"
PASS style.getPropertyValue('border-image-width') is "1"
PASS style.getPropertyValue('border-image-outset') is "2"
PASS style.getPropertyValue('border-image-slice') is "4"
PASS style.getPropertyValue('border-image-width') is "5"
PASS style.getPropertyValue('border-image-outset') is "0"
PASS style.getPropertyValue('border-image-slice') is "1"
PASS style.getPropertyValue('border-image-width') is "2"
PASS style.getPropertyValue('border-image-outset') is "3"
PASS successfullyParsed is true

TEST COMPLETE
